Further split status tests.
