+++ test/mocha/32-status-scaling.js
@@ -31,16 +31,11 @@ describe('issue status scaling', () => {
   for(const suiteName in suiteNames) {
     const suiteInfo = suiteNames[suiteName];
     const {issueOptions, statusOptions, terseIssueOptions} = suiteInfo;
-    if(Array.isArray(suiteInfo.algorithm)) {
-      for(const algorithm of suiteInfo.algorithm) {
-        describeSuite({
-          suiteName, algorithm, issueOptions, statusOptions, terseIssueOptions
-        });
-      }
-    } else {
+    const algorithms = Array.isArray(suiteInfo.algorithm) ?
+      suiteInfo.algorithm : [suiteInfo.algorithm];
+    for(const algorithm of algorithms) {
       describeSuite({
-        suiteName, algorithm: suiteInfo.algorithm, issueOptions,
-        statusOptions, terseIssueOptions
+        suiteName, algorithm, issueOptions, statusOptions, terseIssueOptions
       });
     }
   }
